JPush: Add support for "notification_3rd" payload type
This particular payload type is only document in Chinese here:
https://docs.jiguang.cn/jpush/server/push/rest_api_v3_push#notification_3rd%EF%BC%9A%E8%87%AA%E5%AE%9A%E4%B9%89%E6%B6%88%E6%81%AF%E8%BD%AC%E5%8E%82%E5%95%86%E9%80%9A%E7%9F%A5

The payload is structurally similar to the message payload, but allows
for "eventual" delivery when the device is disconnected for a longer
period of time. Thus this payload should be used for the really important
messages.

An important constraint is that the "message" part of the payload
*must* be present as well.

Reviewed at https://reviews.lunr.nl/r/860/